Simplify the inner manager ledger interface
