+++ base_layer/common_types/Cargo.toml
@@ -19,6 +19,7 @@ newtype-ops = "0.1"
 once_cell = "1.8.0"
 rand = "0.8"
 serde = { version = "1.0.106", features = ["derive"] }
+strum = "0.22"
 thiserror = "1.0.29"
 base64 = "0.21.0"
 blake2 = "0.10"
