+++ base_layer/common_types/src/lib.rs
@@ -32,3 +32,4 @@ pub mod tari_address;
 pub mod transaction;
 mod tx_id;
 pub mod types;
+pub mod wallet_types;
