+++ base_layer/core/src/transactions/key_manager/wrapper.rs
@@ -24,7 +24,10 @@ use std::sync::Arc;
 
 use blake2::Blake2b;
 use digest::consts::U64;
-use tari_common_types::types::{ComAndPubSignature, Commitment, PrivateKey, PublicKey, RangeProof, Signature};
+use tari_common_types::{
+    types::{ComAndPubSignature, Commitment, PrivateKey, PublicKey, RangeProof, Signature},
+    wallet_types::WalletType,
+};
 use tari_comms::types::CommsDHKE;
 use tari_crypto::{hashing::DomainSeparatedHash, ristretto::RistrettoComSig};
 use tari_key_manager::{
@@ -80,12 +83,14 @@ where TBackend: KeyManagerBackend<PublicKey> + 'static
         master_seed: CipherSeed,
         db: KeyManagerDatabase<TBackend, PublicKey>,
         crypto_factories: CryptoFactories,
+        wallet_type: WalletType,
     ) -> Result<Self, KeyManagerServiceError> {
         Ok(TransactionKeyManagerWrapper {
             transaction_key_manager_inner: Arc::new(RwLock::new(TransactionKeyManagerInner::new(
                 master_seed,
                 db,
                 crypto_factories,
+                wallet_type,
             )?)),
         })
     }
