+++ base_layer/wallet/src/lib.rs
@@ -33,7 +33,7 @@ extern crate diesel_migrations;
 mod config;
 pub mod schema;
 pub mod utxo_scanner_service;
-pub use config::{TransactionStage, WalletConfig, WalletType};
+pub use config::{TransactionStage, WalletConfig};
 use tari_contacts::contacts_service::storage::sqlite_db::ContactsServiceSqliteDatabase;
 use tari_core::transactions::key_manager::TransactionKeyManagerWrapper;
 use tari_key_manager::key_manager_service::storage::sqlite_db::KeyManagerSqliteDatabase;
